:art: [Refector] TeamList 구조 수정
- page 에선 setState
- component 에선 view & subscribe
